+++ go/internal/app/keys.go
@@ -47,23 +47,32 @@ import (
 	"golang.org/x/crypto/ssh"
 
 	"github.com/c2FmZQ/sshterm/internal/jsutil"
+	"github.com/c2FmZQ/sshterm/internal/webauthnsk"
 )
 
 func (a *App) generateKey(name, passphrase, idp, typ string, bits int) (*key, error) {
-	pub, priv, err := a.createKey(typ, bits, name)
-	if err != nil {
-		return nil, err
-	}
 	var sshPub ssh.PublicKey
 	var privPEM *pem.Block
-	if k, ok := priv.(*webAuthnKey); ok {
-		pp, err := k.Private(passphrase)
+
+	if typ == "ecdsa-sk" {
+		if bits != 0 && bits != 256 {
+			return nil, fmt.Errorf("invalid key length %d", bits)
+		}
+		sk, err := webauthnsk.Create(name)
+		if err != nil {
+			return nil, fmt.Errorf("webauthnsk.Create: %w", err)
+		}
+		pp, err := sk.MarshalPrivate(passphrase)
+		if err != nil {
+			return nil, fmt.Errorf("sk.MarshalPrivate: %w", err)
+		}
+		privPEM = pp
+		sshPub = sk.PublicKey()
+	} else {
+		pub, priv, err := createKey(typ, bits)
 		if err != nil {
 			return nil, err
 		}
-		privPEM = pp
-		sshPub = k.PublicKey()
-	} else {
 		if sshPub, err = ssh.NewPublicKey(pub); err != nil {
 			return nil, fmt.Errorf("ssh.NewPublicKey: %w", err)
 		}
@@ -90,7 +99,7 @@ func (a *App) keysCommand() *cli.App {
 	return &cli.App{
 		Name:            "keys",
 		Usage:           "Manage user keys and certificates",
-		UsageText:       "keys <list|generate|delete|show|import|import-cert|export>",
+		UsageText:       "keys <list|generate|delete|show|change-pass|import|import-cert|export>",
 		Description:     "The keys command is used to manage user keys and certificates.",
 		HideHelpCommand: true,
 		DefaultCommand:  "list",
@@ -150,7 +159,7 @@ func (a *App) keysCommand() *cli.App {
 						return nil
 					}
 					name := ctx.Args().Get(0)
-					passphrase, err := a.term.ReadPassword("Enter passphrase for private key: ")
+					passphrase, err := a.term.ReadPassword("Enter a passphrase for the private key: ")
 					if err != nil {
 						return fmt.Errorf("ReadPassword: %w", err)
 					}
@@ -224,6 +233,73 @@ func (a *App) keysCommand() *cli.App {
 				},
 			},
 			{
+				Name:      "change-pass",
+				Usage:     "Change a key's passphrase",
+				UsageText: "keys change-pass <name>",
+				Action: func(ctx *cli.Context) error {
+					if ctx.Args().Len() != 1 {
+						cli.ShowSubcommandHelp(ctx)
+						return nil
+					}
+					name := ctx.Args().Get(0)
+					key, exists := a.data.Keys[name]
+					if !exists {
+						return fmt.Errorf("unknown key %q", name)
+					}
+					prompt := func() (string, error) {
+						passphrase, err := a.term.ReadPassword("Enter a NEW passphrase for the private key: ")
+						if err != nil {
+							return "", fmt.Errorf("ReadPassword: %w", err)
+						}
+						passphrase2, err := a.term.ReadPassword("Re-enter the same new passphrase: ")
+						if err != nil {
+							return "", fmt.Errorf("ReadPassword: %w", err)
+						}
+						if passphrase != passphrase2 {
+							return "", fmt.Errorf("passphrase doesn't match")
+						}
+						return passphrase, nil
+					}
+					var privPEM *pem.Block
+					if bytes.HasPrefix(key.Private, []byte("-----BEGIN WEBAUTHN ")) {
+						sk, err := webauthnsk.Unmarshal(key.Private, name, a.term.ReadPassword)
+						if err != nil {
+							return fmt.Errorf("webauthnsk.Unmarshal: %w", err)
+						}
+						pp, err := prompt()
+						if err != nil {
+							return err
+						}
+						if privPEM, err = sk.MarshalPrivate(pp); err != nil {
+							return err
+						}
+					} else {
+						priv, err := key.PrivateKey(a.term.ReadPassword)
+						if err != nil {
+							return fmt.Errorf("%q: %w", name, err)
+						}
+						pp, err := prompt()
+						if err != nil {
+							return err
+						}
+						if pp == "" {
+							if privPEM, err = ssh.MarshalPrivateKey(priv, ""); err != nil {
+								return fmt.Errorf("ssh.MarshalPrivateKey: %w", err)
+							}
+						} else if privPEM, err = ssh.MarshalPrivateKeyWithPassphrase(priv, "", []byte(pp)); err != nil {
+							return fmt.Errorf("ssh.MarshalPrivateKeyWithPassphrase: %w", err)
+						}
+					}
+					key.Private = pem.EncodeToMemory(privPEM)
+					a.data.Keys[name] = key
+					if err := a.saveKeys(true); err != nil {
+						return err
+					}
+					a.term.Printf("Passphrase changed for key %q\n", name)
+					return nil
+				},
+			},
+			{
 				Name:      "import",
 				Usage:     "Import a key",
 				UsageText: "keys import <name>",
@@ -255,21 +331,29 @@ func (a *App) keysCommand() *cli.App {
 						Private: content,
 						errorf:  a.term.Errorf,
 					}
-					priv, err := key.PrivateKey(a.term.ReadPassword)
-					if err != nil {
-						return fmt.Errorf("%q: %w", f.Name, err)
-					}
-					var pub crypto.PublicKey
-					if k, ok := priv.(crypto.Signer); ok {
-						pub = k.Public()
+					if bytes.HasPrefix(content, []byte("-----BEGIN WEBAUTHN ")) {
+						sk, err := webauthnsk.Unmarshal(content, name, a.term.ReadPassword)
+						if err != nil {
+							return fmt.Errorf("webauthnsk.Unmarshal: %w", err)
+						}
+						key.Public = sk.Marshal()
 					} else {
-						return fmt.Errorf("key type %T is not supported", priv)
+						priv, err := key.PrivateKey(a.term.ReadPassword)
+						if err != nil {
+							return fmt.Errorf("%q: %w", f.Name, err)
+						}
+						var pub crypto.PublicKey
+						if k, ok := priv.(crypto.Signer); ok {
+							pub = k.Public()
+						} else {
+							return fmt.Errorf("key type %T is not supported", priv)
+						}
+						sshPub, err := ssh.NewPublicKey(pub)
+						if err != nil {
+							return fmt.Errorf("ssh.NewPublicKey: %w", err)
+						}
+						key.Public = sshPub.Marshal()
 					}
-					sshPub, err := ssh.NewPublicKey(pub)
-					if err != nil {
-						return fmt.Errorf("ssh.NewPublicKey: %w", err)
-					}
-					key.Public = sshPub.Marshal()
 					a.data.Keys[name] = key
 					if err := a.saveKeys(true); err != nil {
 						return err
@@ -375,20 +459,10 @@ func (a *App) keysCommand() *cli.App {
 	}
 }
 
-func (a *App) createKey(t string, b int, name string) (crypto.PublicKey, crypto.PrivateKey, error) {
+func createKey(t string, b int) (crypto.PublicKey, crypto.PrivateKey, error) {
 	switch t {
 	case "ed25519", "":
 		return ed25519.GenerateKey(rand.Reader)
-
-	case "ecdsa-sk":
-		if b != 0 && b != 256 {
-			return nil, nil, fmt.Errorf("invalid key length %d", b)
-		}
-		key, err := createWebAuthnKey(name)
-		if err != nil {
-			return nil, nil, err
-		}
-		return key.PublicKey(), key, nil
 
 	case "ecdsa":
 		if b == 0 {
@@ -494,7 +568,7 @@ func (k *key) isWebAuthn() bool {
 
 func (k *key) sshPublicKey() (ssh.PublicKey, error) {
 	if k.isWebAuthn() {
-		return unmarshalWebAuthnKey(nil, k.Public, k.Name, nil)
+		return webauthnsk.UnmarshalPublic(k.Public)
 	}
 	return ssh.ParsePublicKey(k.Public)
 }
@@ -574,7 +648,7 @@ func (k *key) PrivateKey(rp func(string) (string, error)) (any, error) {
 	if _, ok := err.(*ssh.PassphraseMissingError); !ok || rp == nil {
 		return nil, err
 	}
-	passphrase, err := rp("Enter passphrase for " + k.Name + ": ")
+	passphrase, err := rp("Enter the passphrase for " + k.Name + ": ")
 	if err != nil {
 		return nil, err
 	}
@@ -583,12 +657,12 @@ func (k *key) PrivateKey(rp func(string) (string, error)) (any, error) {
 
 func (k *key) Signer(rp func(string) (string, error)) (ssh.Signer, error) {
 	if k.isWebAuthn() {
-		key, err := unmarshalWebAuthnKey(k.Private, k.Public, k.Name, rp)
+		sk, err := webauthnsk.Unmarshal(k.Private, k.Name, rp)
 		if err != nil {
 			return nil, err
 		}
 		return &dynSigner{
-			Signer: key,
+			Signer: sk,
 			key:    k,
 		}, nil
 	}