+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add WebAuthn keys. The keys can be generated with `keys generate -t ecdsa-sk <name>`. The private key is on a hardware security key, or in a passkey manager.
   Note that the SSH server config needs to have `PubkeyAcceptedAlgorithms +webauthn-sk-ecdsa-sha2-nistp256@openssh.com`.
+* Add `keys change-pass` command to change the passphrase of keys.
 
 ## v0.7.8